Warn if sample names are not unique
- also ensure p5/p7 are all the same when global p5/p7 option is checked
- also refactor main page into SampleSheet component
- resolves #18
